# public/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"error",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "solos_apis",
		"version": "1.0.0",
		"description": "solos apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "BTB"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/superFred/createSuperFred": {
			"post": {
				"operationId": "CreateSuperFred",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Create Super Fred Nft",
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"type": {
										"type": "string"
									},
									"series": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"limit": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"rewardDistribution": {
										"type": "string"
									},
									"rewardSR": {
										"type": "string"
									},
									"dateOfMaturity": {
										"type": "string"
									},
									"maturityAmount": {
										"type": "string"
									},
									"ipfs": {
										"type": "string"
									},
									"image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/superFred/superFredData": {
			"get": {
				"operationId": "SuperFredData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/superFred/superFredDataOfSoloMarketPlace": {
			"get": {
				"operationId": "SuperFredDataOfSoloMarketPlace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"0",
								"1"
							]
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/superFred/buySuperFred": {
			"post": {
				"operationId": "BuySuperFred",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Buy Super Red",
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"walletAddress": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"walletAddress",
									"quantity",
									"price",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/superFred/getUserSuperRed": {
			"get": {
				"operationId": "GetUserSuperRed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/superFred/airdropSfred": {
			"post": {
				"operationId": "AirdropSfred",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "AirDrop superfred",
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "string"
									},
									"userId": {
										"type": "string"
									},
									"sId": {
										"type": "string"
									}
								},
								"required": [
									"userId",
									"sId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/superFred/SFredSellinMarketPlace": {
			"post": {
				"operationId": "SFredSellinMarketPlace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Create Super Fred Nft",
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"walletAddress": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"superFredId": {
										"type": "string"
									}
								},
								"required": [
									"walletAddress",
									"quantity",
									"price",
									"superFredId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/superFred/getSFredInMarketPlace": {
			"get": {
				"operationId": "GetSFredInMarketPlace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"superFreds"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}